perf: shift syncing to workers
